refactor: move function in dedicated file
